Extracted enter on trigger into it's own script
