+++ Assets/EnterableGate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c14b1e0d291e0254fa51a516a4ce12c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
