+++ Assets/Prefabs/Puzzle/ExitGate.prefab
@@ -11,8 +11,8 @@ GameObject:
   - component: {fileID: 8004932041305792291}
   - component: {fileID: 3122119288291986747}
   - component: {fileID: 3624807340573450138}
-  - component: {fileID: 2334755128840050181}
   - component: {fileID: 1364887741390085434}
+  - component: {fileID: 2718195476832982812}
   m_Layer: 0
   m_Name: ExitGate
   m_TagString: Untagged
@@ -104,7 +104,33 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &2334755128840050181
+--- !u!61 &1364887741390085434
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 186770497756555603}
+  m_Enabled: 0
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: -0.0038061142, y: -0.87899697}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0.5, y: 0.5}
+    oldSize: {x: 3.75, y: 3.75}
+    newSize: {x: 3.75, y: 3.75}
+    adaptiveTilingThreshold: 0.5
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 0.9712076, y: 1.9742258}
+  m_EdgeRadius: 0
+--- !u!114 &2718195476832982812
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -113,18 +139,17 @@ MonoBehaviour:
   m_GameObject: {fileID: 186770497756555603}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b2629e2aace8af64e9586aa09969a82a, type: 3}
+  m_Script: {fileID: 11500000, guid: c14b1e0d291e0254fa51a516a4ce12c3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  isActivated: 1
+  isActivated: 0
   logicGate: 0
   invert: 0
-  emitters:
-  - {fileID: 0}
+  emitters: []
   onActivation:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2334755128840050181}
+      - m_Target: {fileID: 2718195476832982812}
         m_TargetAssemblyTypeName: Gate, Assembly-CSharp
         m_MethodName: Toggle
         m_Mode: 1
@@ -151,7 +176,7 @@ MonoBehaviour:
   onDeactivation:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 2334755128840050181}
+      - m_Target: {fileID: 2718195476832982812}
         m_TargetAssemblyTypeName: Gate, Assembly-CSharp
         m_MethodName: Toggle
         m_Mode: 1
@@ -179,6 +204,7 @@ MonoBehaviour:
   panCamera: 0
   showDuration: 1
   enterActionReference: {fileID: 8802194908310570196, guid: 289410948cd3ef44c9fa57193b1dad9c, type: 3}
+  enterThreshold: 0.5
   onEnter:
     m_PersistentCalls:
       m_Calls:
@@ -206,29 +232,3 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
---- !u!61 &1364887741390085434
-BoxCollider2D:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 186770497756555603}
-  m_Enabled: 0
-  m_Density: 1
-  m_Material: {fileID: 0}
-  m_IsTrigger: 1
-  m_UsedByEffector: 0
-  m_UsedByComposite: 0
-  m_Offset: {x: -0.0038061142, y: -0.87899697}
-  m_SpriteTilingProperty:
-    border: {x: 0, y: 0, z: 0, w: 0}
-    pivot: {x: 0.5, y: 0.5}
-    oldSize: {x: 3.75, y: 3.75}
-    newSize: {x: 3.75, y: 3.75}
-    adaptiveTilingThreshold: 0.5
-    drawMode: 0
-    adaptiveTiling: 0
-  m_AutoTiling: 0
-  serializedVersion: 2
-  m_Size: {x: 0.9712076, y: 1.9742258}
-  m_EdgeRadius: 0
