+++ Assets/Scripts/Puzzle/Gate.cs
@@ -11,10 +11,6 @@ public class Gate : SignalReceiver {
 	[SerializeField] private Animator animator;
 	[SerializeField] private bool panCamera;
 	[SerializeField] private float showDuration = 1f;
-	[SerializeField] private InputActionReference enterActionReference;
-	[SerializeField, Range(0.125f, 1f)]
-	private float enterThreshold = 0.5f;
-	[SerializeField] private UnityEvent onEnter;
 
 	private new SnappingCamera camera;
 	private Transform player;
@@ -23,8 +19,6 @@ public class Gate : SignalReceiver {
 	private float cameraTransitionMultiplier = 0.10f;
 	private bool isShowing = false;
 	private bool allowShow = false;
-	private bool canEnter = false;
-	private bool isEntering;
 
 	private void Awake() {
 		GameObject playerObj = GameObject.FindWithTag("Player");
@@ -32,28 +26,6 @@ public class Gate : SignalReceiver {
 		playerController = playerObj.GetComponent<PlayerController>();
 		camera = Camera.main.GetComponent<SnappingCamera>();
 		cameraTransitionDuration = camera.TransitionDuration;
-	}
-
-	private void OnTriggerEnter2D(Collider2D other) {
-		if (other.attachedRigidbody.CompareTag("Player") && !other.isTrigger)
-			canEnter = true;
-	}
-
-	private void OnTriggerExit2D(Collider2D other) {
-		if (other.attachedRigidbody.CompareTag("Player") && !other.isTrigger)
-			canEnter = false;
-	}
-
-	private void OnEnable() {
-		if (enterActionReference == null) return;
-
-		enterActionReference.action.performed += OnEnterInput;
-	}
-
-	private void OnDisable() {
-		if (enterActionReference == null) return;
-
-		enterActionReference.action.performed -= OnEnterInput;
 	}
 
 	public void Toggle() {
@@ -87,20 +59,5 @@ public class Gate : SignalReceiver {
 		playerController.AllowControls = true;
 		Time.timeScale = 1;
 		isShowing = false;
-	}
-
-	private void OnEnterInput(InputAction.CallbackContext ctx) {
-		if (!canEnter) return;
-
-		Vector2 moveInput = ctx.ReadValue<Vector2>();
-		if (moveInput.y > enterThreshold) {
-			if (!isEntering) {
-				isEntering = true;
-				onEnter.Invoke();
-			}
-		}
-		else {
-			isEntering = false;
-		}
 	}
 }
